fix test and move mock
