+++ cmd/axelard/cmd/vald/multisig/keygen_test.go
@@ -2,6 +2,7 @@ package multisig_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 	"time"
 
@@ -9,13 +10,12 @@ import (
 	sdkclient "github.com/cosmos/cosmos-sdk/client"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/assert"
-	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/log"
 	"google.golang.org/grpc"
 
 	broadcastmock "github.com/axelarnetwork/axelar-core/cmd/axelard/cmd/vald/broadcast/mock"
 	"github.com/axelarnetwork/axelar-core/cmd/axelard/cmd/vald/multisig"
-	"github.com/axelarnetwork/axelar-core/cmd/axelard/cmd/vald/tss/rpc/mock"
+	"github.com/axelarnetwork/axelar-core/cmd/axelard/cmd/vald/multisig/mock"
 	"github.com/axelarnetwork/axelar-core/testutils/rand"
 	"github.com/axelarnetwork/axelar-core/x/multisig/exported/testutils"
 	"github.com/axelarnetwork/axelar-core/x/multisig/types"
@@ -29,14 +29,14 @@ func TestMgr_ProcessKeygenStarted(t *testing.T) {
 	var (
 		mgr         *multisig.Mgr
 		participant sdk.ValAddress
-		client      *mock.MultiSigClientMock
+		client      *mock.ClientMock
 		broadcaster *broadcastmock.BroadcasterMock
 
-		e abci.Event
+		event *types.KeygenStarted
 	)
 
 	givenMgr := Given("the multisig manager", func() {
-		client = &mock.MultiSigClientMock{}
+		client = &mock.ClientMock{}
 		broadcaster = &broadcastmock.BroadcasterMock{}
 		participant = rand.ValAddr()
 
@@ -52,13 +52,10 @@ func TestMgr_ProcessKeygenStarted(t *testing.T) {
 
 	givenMgr.
 		When("is not part of the listed participants", func() {
-			event := funcs.Must(
-				sdk.TypedEventToEvent(types.NewKeygenStarted(testutils.KeyID(), slices.Expand(func(int) sdk.ValAddress { return rand.ValAddr() }, 10))),
-			)
-			e = abci.Event{Type: event.Type, Attributes: event.Attributes}
+			event = types.NewKeygenStarted(testutils.KeyID(), slices.Expand(func(int) sdk.ValAddress { return rand.ValAddr() }, 10))
 		}).
 		Then("should ignore", func(t *testing.T) {
-			err := mgr.ProcessKeygenStarted(e)
+			err := mgr.ProcessKeygenStarted(event)
 
 			assert.NoError(t, err)
 		}).
@@ -66,10 +63,7 @@ func TestMgr_ProcessKeygenStarted(t *testing.T) {
 
 	givenMgr.
 		When("is part of the listed participants", func() {
-			event := funcs.Must(
-				sdk.TypedEventToEvent(types.NewKeygenStarted(testutils.KeyID(), []sdk.ValAddress{rand.ValAddr(), participant, rand.ValAddr()})),
-			)
-			e = abci.Event{Type: event.Type, Attributes: event.Attributes}
+			event = types.NewKeygenStarted(testutils.KeyID(), []sdk.ValAddress{rand.ValAddr(), participant, rand.ValAddr()})
 		}).
 		Then("should ignore", func(t *testing.T) {
 			sk := funcs.Must(btcec.NewPrivateKey(btcec.S256()))
@@ -79,15 +73,25 @@ func TestMgr_ProcessKeygenStarted(t *testing.T) {
 			client.SignFunc = func(_ context.Context, in *tofnd.SignRequest, _ ...grpc.CallOption) (*tofnd.SignResponse, error) {
 				return &tofnd.SignResponse{SignResponse: &tofnd.SignResponse_Signature{Signature: funcs.Must(sk.Sign(in.MsgToSign)).Serialize()}}, nil
 			}
-			broadcaster.BroadcastFunc = func(ctx context.Context, msgs ...sdk.Msg) (*sdk.TxResponse, error) { return &sdk.TxResponse{}, nil }
+			broadcaster.BroadcastFunc = func(ctx context.Context, msgs ...sdk.Msg) (*sdk.TxResponse, error) {
+				for _, msg := range msgs {
+					if _, ok := msg.(*types.SubmitPubKeyRequest); !ok {
+						return nil, fmt.Errorf("unexpected type of msg %T received", msg)
+					}
 
-			err := mgr.ProcessKeygenStarted(e)
+					if err := msg.ValidateBasic(); err != nil {
+						return nil, err
+					}
+				}
+
+				return &sdk.TxResponse{}, nil
+			}
+
+			err := mgr.ProcessKeygenStarted(event)
 			assert.NoError(t, err)
 
 			assert.Len(t, broadcaster.BroadcastCalls(), 1)
 			assert.Len(t, broadcaster.BroadcastCalls()[0].Msgs, 1)
-			assert.IsType(t, broadcaster.BroadcastCalls()[0].Msgs[0], &types.SubmitPubKeyRequest{})
-			assert.NoError(t, broadcaster.BroadcastCalls()[0].Msgs[0].(*types.SubmitPubKeyRequest).ValidateBasic())
 		}).
 		Run(t)
 }
